vaporbat/requirements.txt
PyCrypto
imaplib2
